refactor(be): 고객 분석 요청 시 data : [ {...}, ... ]의 형태로 보내도록 형식 수정
