+++ backend/src/main/java/com/example/moki_campaign/domain/customer/service/CustomerServiceImpl.java
@@ -16,12 +16,10 @@ import com.example.moki_campaign.domain.visit.entity.DailyVisit;
 import com.example.moki_campaign.domain.visit.repository.DailyVisitRepository;
 import com.example.moki_campaign.global.exception.common.BusinessException;
 import com.example.moki_campaign.global.exception.common.ErrorCode;
-import com.example.moki_campaign.global.util.DateRangeCalculator;
-import com.example.moki_campaign.global.util.DateRangeCalculator.DateRange;
 import com.example.moki_campaign.infra.ai.client.AiClient;
-import com.example.moki_campaign.infra.ai.dto.AiCustomerDataInputDto;
-import com.example.moki_campaign.infra.ai.dto.AiCustomerDataOutputDto;
-import com.example.moki_campaign.infra.ai.dto.AiCustomerDataResponseDto;
+import com.example.moki_campaign.infra.ai.dto.request.AiCustomerDataInputDto;
+import com.example.moki_campaign.infra.ai.dto.response.AiCustomerDataOutputDto;
+import com.example.moki_campaign.infra.ai.dto.response.AiCustomerDataResponseDto;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
