+++ backend/src/main/java/com/example/moki_campaign/infra/ai/dto/request/AiAnalysisRequestDto.java
@@ -1,0 +1,8 @@
+package com.example.moki_campaign.infra.ai.dto.request;
+
+import java.util.List;
+
+public record AiAnalysisRequestDto(
+        List<AiCustomerDataInputDto> data
+) {
+}
