+++ backend/src/main/java/com/example/moki_campaign/infra/ai/dto/request/AiCustomerDataInputDto.java
@@ -1,4 +1,4 @@
-package com.example.moki_campaign.infra.ai.dto;
+package com.example.moki_campaign.infra.ai.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
