+++ backend/src/main/java/com/example/moki_campaign/infra/ai/dto/response/AiCustomerDataOutputDto.java
@@ -1,4 +1,4 @@
-package com.example.moki_campaign.infra.ai.dto;
+package com.example.moki_campaign.infra.ai.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
