+++ backend/src/test/java/com/example/moki_campaign/customer/CustomerServiceImplTest.java
@@ -16,9 +16,9 @@ import com.example.moki_campaign.domain.visit.repository.DailyVisitRepository;
 import com.example.moki_campaign.global.exception.common.BusinessException;
 import com.example.moki_campaign.global.exception.common.ErrorCode;
 import com.example.moki_campaign.infra.ai.client.AiClient;
-import com.example.moki_campaign.infra.ai.dto.AiCustomerDataInputDto;
-import com.example.moki_campaign.infra.ai.dto.AiCustomerDataOutputDto;
-import com.example.moki_campaign.infra.ai.dto.AiCustomerDataResponseDto;
+import com.example.moki_campaign.infra.ai.dto.request.AiCustomerDataInputDto;
+import com.example.moki_campaign.infra.ai.dto.response.AiCustomerDataOutputDto;
+import com.example.moki_campaign.infra.ai.dto.response.AiCustomerDataResponseDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
